[refactor] 더미 클래스를 실제 User 도메인으로 교체
