+++ src/main/java/com/prgrms/rg/domain/notification/application/impl/NotificationPushService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.event.TransactionalEventListener;
 import com.prgrms.rg.domain.notification.application.NotificationService;
 import com.prgrms.rg.domain.notification.model.Notification;
 import com.prgrms.rg.domain.notification.model.sse.SseNotificationSender;
+import com.prgrms.rg.domain.ridingpost.model.RidingPost;
 import com.prgrms.rg.domain.ridingpost.model.RidingPostFinder;
-import com.prgrms.rg.domain.ridingpost.model.dummy.RidingPost;
 import com.prgrms.rg.domain.ridingpost.model.event.RidingJoinEvent;
 import com.prgrms.rg.domain.user.model.User;
 
@@ -32,7 +32,7 @@ public class NotificationPushService {
 	@Transactional(propagation = Propagation.REQUIRES_NEW)
 	public void handleRidingJoinEvent(RidingJoinEvent event) {
 		RidingPost post = postFinder.find(event.getPostId());
-		User host = post.getHost();
+		User host = post.getLeader();
 		Notification notification = notificationService.createRidingJoinNotification(host.getId(), post.getId());
 		notificationSender.sendNotification(host.getId(), notification, RIDING_JOIN_EVENT);
 	}
