+++ src/main/java/com/prgrms/rg/domain/ridingpost/model/RidingParticipantRepository.java
@@ -1,6 +1,5 @@
 package com.prgrms.rg.domain.ridingpost.model;
 
-import com.prgrms.rg.domain.ridingpost.model.dummy.RidingPost;
 import com.prgrms.rg.domain.user.model.User;
 
 public interface RidingParticipantRepository {
