+++ src/main/java/com/prgrms/rg/domain/ridingpost/model/RidingPostRepository.java
@@ -2,7 +2,6 @@ package com.prgrms.rg.domain.ridingpost.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.prgrms.rg.domain.ridingpost.model.dummy.RidingPost;
 
 public interface RidingPostRepository extends JpaRepository<RidingPost, Long> {
 }
